+++ src/vs/workbench/browser/parts/notifications/notificationsCommands.ts
@@ -17,7 +17,8 @@ import { NotificationFocusedContext, NotificationsCenterVisibleContext, Notifica
 import { CONTEXT_ACCESSIBILITY_MODE_ENABLED } from 'vs/platform/accessibility/common/accessibility';
 import { INotificationService, NotificationPriority } from 'vs/platform/notification/common/notification';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
-import { NotificationActionRunner } from 'vs/workbench/browser/parts/notifications/notificationsActions';
+import { ActionRunner, IAction, WorkbenchActionExecutedEvent, WorkbenchActionExecutedClassification } from 'vs/base/common/actions';
+import { hash } from 'vs/base/common/hash';
 
 // Center
 export const SHOW_NOTIFICATIONS_CENTER = 'notifications.showList';
@@ -275,4 +276,53 @@ export function registerNotificationCommands(center: INotificationsCenterControl
 	MenuRegistry.appendMenuItem(MenuId.CommandPalette, { command: { id: CLEAR_ALL_NOTIFICATIONS, title: { value: localize('clearAllNotifications', "Clear All Notifications"), original: 'Clear All Notifications' }, category } });
 	MenuRegistry.appendMenuItem(MenuId.CommandPalette, { command: { id: TOGGLE_DO_NOT_DISTURB_MODE, title: { value: localize('toggleDoNotDisturbMode', "Toggle Do Not Disturb Mode"), original: 'Toggle Do Not Disturb Mode' }, category } });
 	MenuRegistry.appendMenuItem(MenuId.CommandPalette, { command: { id: FOCUS_NOTIFICATION_TOAST, title: { value: localize('focusNotificationToasts', "Focus Notification Toast"), original: 'Focus Notification Toast' }, category }, when: NotificationsToastsVisibleContext });
+}
+
+
+interface NotificationActionMetrics {
+	readonly id: string;
+	readonly actionLabel: string;
+	readonly source: string;
+	readonly silent: boolean;
+}
+
+type NotificationActionMetricsClassification = {
+	id: { classification: 'SystemMetaData'; purpose: 'FeatureInsight'; comment: 'The identifier of the action that was run from a notification.' };
+	actionLabel: { classification: 'SystemMetaData'; purpose: 'FeatureInsight'; comment: 'The label of the action that was run from a notification.' };
+	source: { classification: 'SystemMetaData'; purpose: 'FeatureInsight'; comment: 'The source of the notification where an action was run.' };
+	silent: { classification: 'SystemMetaData'; purpose: 'FeatureInsight'; comment: 'Whether the notification where an action was run is silent or not.' };
+	owner: 'bpasero';
+	comment: 'Tracks when actions are fired from notifcations and how they were fired.';
+};
+
+export class NotificationActionRunner extends ActionRunner {
+
+	constructor(
+		@ITelemetryService private readonly telemetryService: ITelemetryService,
+		@INotificationService private readonly notificationService: INotificationService
+	) {
+		super();
+	}
+
+	protected override async runAction(action: IAction, context: unknown): Promise<void> {
+		this.telemetryService.publicLog2<WorkbenchActionExecutedEvent, WorkbenchActionExecutedClassification>('workbenchActionExecuted', { id: action.id, from: 'message' });
+
+		if (isNotificationViewItem(context)) {
+			// Log some additional telemetry specifically for actions
+			// that are triggered from within notifications.
+			this.telemetryService.publicLog2<NotificationActionMetrics, NotificationActionMetricsClassification>('notification:actionExecuted', {
+				id: hash(context.message.original.toString()).toString(),
+				actionLabel: action.label,
+				source: context.sourceId || 'core',
+				silent: context.priority === NotificationPriority.SILENT
+			});
+		}
+
+		// Run and make sure to notify on any error again
+		try {
+			await super.runAction(action, context);
+		} catch (error) {
+			this.notificationService.error(error);
+		}
+	}
 }
